error 1000 in stream offline
the server always returns 200  code when call. i use this for see if the server is online but is conflicting with twitch.
i moved it to "proxystatus" on header.

but it need update the extension app for do it work, otherwise the app could power off the adblock when received a 404 return from server.